feat: replace RateTranslateButton with inline translation button in RateItem component
