+++ src/components/Rate/index.ts
@@ -5,7 +5,6 @@ export { RateItemSkeleton } from "./RateItemSkeleton";
 export { RateForm } from "./RateForm";
 export { RateEditDialog } from "./RateEditDialog";
 export { RateDeleteDialog } from "./RateDeleteDialog";
-export { RateTranslateButton } from "./rate-translate-button";
 
 export type { RateFormValues } from "./RateForm";
